Reconfigure such classes as Student and Person into pointer class.
So, the last one, it's Teacher class, is ruins my plans, because i have forgot that it's should contain objects of Subject class, and I don't know, should I remake in into pointer class as the others or leave it be? Maybe.... Well, I was going to turn every classes of Person family into pointer class, thats mean I have to eventually turn into pointer class. Damn, but it's nothing, I can hande it. It's not requires any changes in Subjects class, which is good.