Create a home page of Interface
Uhuh, so is suppose to look approximetaly like this. As my classmate said, It could looks like program we were making in the previous semester.

Step by step and I will be able to finish this project in no time. So please, Mark, don't give up and look forward only to success. I believe in you