Create loadStudents() function in Interface class
This fuction basically just load data about students in the program. But now I have a problem with marks, since I was planning to create a separate file called "marks.txt" which supposed to hold all grades students have. But the point is class Student have an object marks in it, which is a problem for meeee. Maybe not for anyone else, but now I'm planning to add some upgrades to function loadStudents().
As you might notice, I also created class saveStudents, but it's another story, since I guess it will requre much more of my time. Sadly, I don't have much time(((