Some little changes to createStudent() and loadStudent() funcions.
Now it's looks more fashionita and buonita and stuff and also it can now import Addresses with spaces! It wasn't so hard to implicate it, sp there is nothing to be proud of(. Now I wan't to allow users to look detailed info about students. In fact, there will be usual info but with grades. At this point, I should make grid of grades first, but it will be such a pain in the ass today. Maybe tomorrow I'll try to implicate it into my project yaaaay