Create sortStudents() function
Finally! It was so hard to found any normal code, that will not crash any processes, and you know what? The reason why I wasted so much time and energy is that I didn't initialized move and copy constructors in Student class, which is the daughter class of Person class. I was thinking the whole time that if i will initialize those constructors in parent class, I will have no need no do the same thing in it's daughter class, considering that the class Student don't have atributes different from Person. But I was fking wrong! I wasted more that an hour to solve the problem, that's insane! But at least, I did it, and tbh this function is very funny to play with. Like, it's hella entertaining to see how it sorts by different attributes of Student class.
In general, now we can consider that the half of my program is done. Now I have to do other two main functions, such as gradesView() and teachersView(), complete with studentsView(), because despite all functions that allow me to create students' info, erase it and view, I can't manipulate with it's attributes and marks. To be more precisely, I can't edit the info of students. But since I already know, that it will be no easy work, I will leave it for later.
Eventually, all of that work could have been much easier if I would have started it in the summer, or should I say 2 freaking month ago! Oh, wait, I did started long time ago, but some time passed and I completely forgot about it. Shame on me, now I have to pass a lot of classes just to complete this project. That's crazy ngl. I'm sure I will remember this event for a long time(