Fix eraseMarks() function
While doing things with gradesView, I have noticed that eraseMarks() function isn't work properly. And after all, it wasn't hard at all to fix it, since as a so called "DoomedStudentId" I've used wrong value — instead, I've used "decition" which decide if this function will be caller or not. Of course, the standard value of this "decition" was '1', so the fuction was deleting only grades of the first Id. Stupid mistake, but I'm still learning!