MyJournal 0.3.1 / Update createStudent() and gradeView() funcs
I decided that now we can consider it as a new modul of my project, since from this moment we will focus more attention on grade functions.  So now it's new version, yaaay, it's already 0.3.1!!! I don't know what sub-version will be final before full version (I mean 1.0.1), but ummm let's predict it! I feeell! I feeeeeel, that it will be 0.5.5! Or somewhere around this number. We will see it in Monday, I guess, since I'm totally sure that I will not make it 'till the end of Sunday((((