Create sortGrades() function
I was trying so hard to fix this hella sorting function because it was kinda complicated. At first I was planning to code it to sort by idSubject and then by idGrades, but it was working really badly, ngl. After some atemps I gave up and started to making up another sorting system. And, after not so long, I was able to find the solution. It's pretty easy, first it sorts by idMarks and then by idSubjects. Insanely easy, I know, but it wasn't so obvious from the start!