Reduced and optimized several functions
After a while I was managed to reduce amount of function of Interface class and now it looks more or less nicer, yaaaay. But unfortunately it cost me a lot of time, what is horrible. I need to hurry up and spend less time on such things as this. [sad miserable emoji]