MyJournal 1.0.1
Yaaaaaasssssssss!!!!!!! Finally!!! You see that? I have finally done it! Now it's pretty suitable to be used in work he was designed to.
Unfortunately tho, I didn't finished my documentation to this code, so I might be in a problem right now, but it's not like I care much about is, I made my own program! Let me be proud for a second
I have done it, but it's doesn't mean that I am giving up on this project. At the free time, I will upgrade it more and more until I'll reach my limits, yaay